requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
dnspython==2.4.2

# Security (JWT & Hashing)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP Client (Para comunicación entre servicios)
httpx==0.25.2
